core/pgxp: fix C++<20 UB on signed left shift in TruncateVertexPosition
TruncateVertexPosition emulates the GPU hardware behaviour of
truncating vertex positions to 11-bit signed (the same range as
the BitField<u32, s32, X, 11> declarations in gpu_types.h). The
implementation uses the classic sign-extend-via-shift idiom:

  static_cast<s16>(int_part << 5) >> 5

with shift = 5 to land the field's high bit at s16's sign bit
position (16 - 11 = 5). The arithmetic right shift then replicates
the sign bit. The integer-part of a vertex coordinate routinely
sits in the negative half of the 11-bit signed range
(int_part = -1024..-1 for any vertex with X or Y below zero),
which means int_part << 5 is a signed left shift of a negative
value: undefined behaviour in C++<20, well-defined two's-complement
since C++20.

Fix: perform the left shift in the unsigned domain, then cast to
s16 for the arithmetic right shift. Same idiom as the recently-
fixed BitField::GetValue() signed branch, the mdec.cpp MDEC
coefficient sign-extension, and TruncateGPUVertexPosition in
gpu_types.h.

  static_cast<s16>(static_cast<u32>(int_part) << 5) >> 5

Observable behaviour: zero change. The fix is codegen-byte-
identical at -O2 g++ x86_64 - both forms emit:

  sall $5 ; sarw $5 ; cwtl

GCC's UB choice for the old form happens to match the C++20-
defined behaviour, so all 11 boundary-test cases (zero, edges of
the 11-bit range, just-past-edges that wrap, large positive and
negative values that exit the field) give identical floats from
both forms.

The fix removes the UB so the source stays well-defined under
future compiler changes and across compilers.

+++ src/core/pgxp.cpp
@@ -500,7 +500,11 @@ static ALWAYS_INLINE_RELEASE float TruncateVertexPosition(float p)
 {
   const s32 int_part = static_cast<s32>(p);
   const float int_part_f = static_cast<float>(int_part);
-  return static_cast<float>(static_cast<s16>(int_part << 5) >> 5) + (p - int_part_f);
+  // Sign-extend the low 11 bits via shift idiom. Do the left shift in
+  // unsigned space to avoid C++<20 UB on negative int_part (the GPU
+  // truncates positions to an 11-bit signed range, so int_part is
+  // routinely negative).
+  return static_cast<float>(static_cast<s16>(static_cast<u32>(int_part) << 5) >> 5) + (p - int_part_f);
 }
 
 static ALWAYS_INLINE_RELEASE bool IsWithinTolerance(float precise_x, float precise_y, int int_x, int int_y)
